Attribute-Handlers: Swap out base for parent in test files.
Committer's note: Add demo/MyClass.pm to list of
files whose $VERSION numbers do not have to be incremented when modified.
Since this file is found in a 'demo/' directory, it does not have to be
treated as if it were a real Perl library which happens to be shipping
with core.

For: RT #119365

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>